+++ libs/coin-modules/coin-evm/src/__tests__/coin-tester/scenarii/blast.ts
@@ -11,22 +11,18 @@ import { buildAccountBridge, buildCurrencyBridge } from "../../../bridge/js";
 import { getCoinConfig, setCoinConfig } from "../../../config";
 import { Transaction as EvmTransaction } from "../../../types";
 import { makeAccount } from "../../fixtures/common.fixtures";
-import { callMyDealer, scroll, VITALIK } from "../helpers";
-import { defaultNanoApp } from "../scenarios.test";
+import { blast, callMyDealer, VITALIK } from "../helpers";
+import { defaultNanoApp } from "../scenarii.test";
 import { killAnvil, spawnAnvil } from "../anvil";
 import resolver from "../../../hw-getAddress";
 
-type ScrollScenarioTransaction = ScenarioTransaction<EvmTransaction, Account>;
+type BlastScenarioTransaction = ScenarioTransaction<EvmTransaction, Account>;
 
 // getTokenById will only work after the currency has been preloaded
-const TOKEN_ID = "scroll/erc20/usd_coin";
+const TOKEN_ID = "blast/erc20/magic_internet_money";
 
-const makeScenarioTransactions = ({
-  address,
-}: {
-  address: string;
-}): ScrollScenarioTransaction[] => {
-  const scenarioSendEthTransaction: ScrollScenarioTransaction = {
+const makeScenarioTransactions = ({ address }: { address: string }): BlastScenarioTransaction[] => {
+  const scenarioSendEthTransaction: BlastScenarioTransaction = {
     name: "Send 1 ETH",
     amount: new BigNumber(1e18),
     recipient: VITALIK,
@@ -41,14 +37,14 @@ const makeScenarioTransactions = ({
     },
   };
 
-  const USDC_ON_SCROLL = getTokenById("scroll/erc20/usd_coin");
-  const scenarioSendUSDCTransaction: ScrollScenarioTransaction = {
-    name: "Send USDC",
+  const MIM_ON_BLAST = getTokenById(TOKEN_ID);
+  const scenarioSendMIMTransaction: BlastScenarioTransaction = {
+    name: "Send 80 MIM",
     amount: new BigNumber(
-      ethers.utils.parseUnits("80", USDC_ON_SCROLL.units[0].magnitude).toString(),
+      ethers.utils.parseUnits("80", MIM_ON_BLAST.units[0].magnitude).toString(),
     ),
     recipient: VITALIK,
-    subAccountId: encodeTokenAccountId(`js:2:scroll:${address}:`, USDC_ON_SCROLL),
+    subAccountId: encodeTokenAccountId(`js:2:blast:${address}:`, MIM_ON_BLAST),
     expect: (previousAccount, currentAccount) => {
       const [latestOperation] = currentAccount.operations;
       expect(currentAccount.operations.length - previousAccount.operations.length).toBe(1);
@@ -56,23 +52,23 @@ const makeScenarioTransactions = ({
       expect(latestOperation.value.toFixed()).toBe(latestOperation.fee.toFixed());
       expect(latestOperation.subOperations?.[0].type).toBe("OUT");
       expect(latestOperation.subOperations?.[0].value.toFixed()).toBe(
-        ethers.utils.parseUnits("80", USDC_ON_SCROLL.units[0].magnitude).toString(),
+        ethers.utils.parseUnits("80", MIM_ON_BLAST.units[0].magnitude).toString(),
       );
       expect(currentAccount.subAccounts?.[0].balance.toFixed()).toBe(
-        ethers.utils.parseUnits("20", USDC_ON_SCROLL.units[0].magnitude).toString(),
+        ethers.utils.parseUnits("20", MIM_ON_BLAST.units[0].magnitude).toString(),
       );
     },
   };
 
-  return [scenarioSendEthTransaction, scenarioSendUSDCTransaction];
+  return [scenarioSendEthTransaction, scenarioSendMIMTransaction];
 };
 
-export const scenarioScroll: Scenario<EvmTransaction, Account> = {
-  name: "Ledger Live Basic Scroll Transactions",
+export const scenarioBlast: Scenario<EvmTransaction, Account> = {
+  name: "Ledger Live Basic Blast Transactions",
   setup: async () => {
     const [{ transport, getOnSpeculosConfirmation }] = await Promise.all([
       spawnSpeculos(`/${defaultNanoApp.firmware}/Ethereum/app_${defaultNanoApp.version}.elf`),
-      spawnAnvil("https://rpc.scroll.io"),
+      spawnAnvil("https://rpc.blast.io"),
     ]);
 
     const provider = new providers.StaticJsonRpcProvider("http://127.0.0.1:8545");
@@ -93,31 +89,31 @@ export const scenarioScroll: Scenario<EvmTransaction, Account> = {
         },
         explorer: {
           type: "etherscan",
-          uri: "https://api.scrollscan.com/api",
+          uri: "https://api.blastscan.io/api",
         },
       },
     }));
-    initMswHandlers(getCoinConfig(scroll).info);
+    initMswHandlers(getCoinConfig(blast).info);
 
     const onSignerConfirmation = getOnSpeculosConfirmation();
     const currencyBridge = buildCurrencyBridge(signerContext);
-    await currencyBridge.preload(scroll);
+    await currencyBridge.preload(blast);
     const accountBridge = buildAccountBridge(signerContext);
     const getAddress = resolver(signerContext);
     const { address } = await getAddress("", {
       path: "44'/60'/0'/0/0",
-      currency: scroll,
+      currency: blast,
       derivationMode: "",
     });
 
-    const scenarioAccount = makeAccount(address, scroll);
+    const scenarioAccount = makeAccount(address, blast);
 
-    const USDC_ON_SCROLL = getTokenById(TOKEN_ID);
+    const MIM_ON_BLAST = getTokenById(TOKEN_ID);
     await callMyDealer({
       provider,
-      drug: USDC_ON_SCROLL,
+      drug: MIM_ON_BLAST,
       junkie: address,
-      dose: ethers.utils.parseUnits("100", USDC_ON_SCROLL.units[0].magnitude),
+      dose: ethers.utils.parseUnits("100", MIM_ON_BLAST.units[0].magnitude),
     });
 
     return {
@@ -133,18 +129,18 @@ export const scenarioScroll: Scenario<EvmTransaction, Account> = {
     await indexBlocks();
   },
   beforeAll: account => {
-    const USDC_ON_SCROLL = getTokenById(TOKEN_ID);
+    const MIM_ON_BLAST = getTokenById(TOKEN_ID);
     expect(account.balance.toFixed()).toBe(ethers.utils.parseEther("10000").toString());
     expect(account.subAccounts?.[0]?.type).toBe("TokenAccount");
     expect(account.subAccounts?.[0]?.balance?.toFixed()).toBe(
-      ethers.utils.parseUnits("100", USDC_ON_SCROLL.units[0].magnitude).toString(),
+      ethers.utils.parseUnits("100", MIM_ON_BLAST.units[0].magnitude).toString(),
     );
   },
   afterAll: account => {
-    const USDC_ON_SCROLL = getTokenById(TOKEN_ID);
+    const MIM_ON_BLAST = getTokenById(TOKEN_ID);
     expect(account.subAccounts?.length).toBe(1);
     expect(account.subAccounts?.[0].balance.toFixed()).toBe(
-      ethers.utils.parseUnits("20", USDC_ON_SCROLL.units[0].magnitude).toString(),
+      ethers.utils.parseUnits("20", MIM_ON_BLAST.units[0].magnitude).toString(),
     );
     expect(account.operations.length).toBe(3);
   },
