+++ libs/coin-modules/coin-evm/src/__tests__/coin-tester/scenarii/scroll.ts
@@ -11,18 +11,22 @@ import { buildAccountBridge, buildCurrencyBridge } from "../../../bridge/js";
 import { getCoinConfig, setCoinConfig } from "../../../config";
 import { Transaction as EvmTransaction } from "../../../types";
 import { makeAccount } from "../../fixtures/common.fixtures";
-import { blast, callMyDealer, VITALIK } from "../helpers";
-import { defaultNanoApp } from "../scenarios.test";
+import { callMyDealer, scroll, VITALIK } from "../helpers";
+import { defaultNanoApp } from "../scenarii.test";
 import { killAnvil, spawnAnvil } from "../anvil";
 import resolver from "../../../hw-getAddress";
 
-type BlastScenarioTransaction = ScenarioTransaction<EvmTransaction, Account>;
+type ScrollScenarioTransaction = ScenarioTransaction<EvmTransaction, Account>;
 
 // getTokenById will only work after the currency has been preloaded
-const TOKEN_ID = "blast/erc20/magic_internet_money";
+const TOKEN_ID = "scroll/erc20/usd_coin";
 
-const makeScenarioTransactions = ({ address }: { address: string }): BlastScenarioTransaction[] => {
-  const scenarioSendEthTransaction: BlastScenarioTransaction = {
+const makeScenarioTransactions = ({
+  address,
+}: {
+  address: string;
+}): ScrollScenarioTransaction[] => {
+  const scenarioSendEthTransaction: ScrollScenarioTransaction = {
     name: "Send 1 ETH",
     amount: new BigNumber(1e18),
     recipient: VITALIK,
@@ -37,14 +41,14 @@ const makeScenarioTransactions = ({ address }: { address: string }): BlastScenar
     },
   };
 
-  const MIM_ON_BLAST = getTokenById(TOKEN_ID);
-  const scenarioSendMIMTransaction: BlastScenarioTransaction = {
-    name: "Send 80 MIM",
+  const USDC_ON_SCROLL = getTokenById("scroll/erc20/usd_coin");
+  const scenarioSendUSDCTransaction: ScrollScenarioTransaction = {
+    name: "Send USDC",
     amount: new BigNumber(
-      ethers.utils.parseUnits("80", MIM_ON_BLAST.units[0].magnitude).toString(),
+      ethers.utils.parseUnits("80", USDC_ON_SCROLL.units[0].magnitude).toString(),
     ),
     recipient: VITALIK,
-    subAccountId: encodeTokenAccountId(`js:2:blast:${address}:`, MIM_ON_BLAST),
+    subAccountId: encodeTokenAccountId(`js:2:scroll:${address}:`, USDC_ON_SCROLL),
     expect: (previousAccount, currentAccount) => {
       const [latestOperation] = currentAccount.operations;
       expect(currentAccount.operations.length - previousAccount.operations.length).toBe(1);
@@ -52,23 +56,23 @@ const makeScenarioTransactions = ({ address }: { address: string }): BlastScenar
       expect(latestOperation.value.toFixed()).toBe(latestOperation.fee.toFixed());
       expect(latestOperation.subOperations?.[0].type).toBe("OUT");
       expect(latestOperation.subOperations?.[0].value.toFixed()).toBe(
-        ethers.utils.parseUnits("80", MIM_ON_BLAST.units[0].magnitude).toString(),
+        ethers.utils.parseUnits("80", USDC_ON_SCROLL.units[0].magnitude).toString(),
       );
       expect(currentAccount.subAccounts?.[0].balance.toFixed()).toBe(
-        ethers.utils.parseUnits("20", MIM_ON_BLAST.units[0].magnitude).toString(),
+        ethers.utils.parseUnits("20", USDC_ON_SCROLL.units[0].magnitude).toString(),
       );
     },
   };
 
-  return [scenarioSendEthTransaction, scenarioSendMIMTransaction];
+  return [scenarioSendEthTransaction, scenarioSendUSDCTransaction];
 };
 
-export const scenarioBlast: Scenario<EvmTransaction, Account> = {
-  name: "Ledger Live Basic Blast Transactions",
+export const scenarioScroll: Scenario<EvmTransaction, Account> = {
+  name: "Ledger Live Basic Scroll Transactions",
   setup: async () => {
     const [{ transport, getOnSpeculosConfirmation }] = await Promise.all([
       spawnSpeculos(`/${defaultNanoApp.firmware}/Ethereum/app_${defaultNanoApp.version}.elf`),
-      spawnAnvil("https://rpc.blast.io"),
+      spawnAnvil("https://rpc.scroll.io"),
     ]);
 
     const provider = new providers.StaticJsonRpcProvider("http://127.0.0.1:8545");
@@ -89,31 +93,31 @@ export const scenarioBlast: Scenario<EvmTransaction, Account> = {
         },
         explorer: {
           type: "etherscan",
-          uri: "https://api.blastscan.io/api",
+          uri: "https://api.scrollscan.com/api",
         },
       },
     }));
-    initMswHandlers(getCoinConfig(blast).info);
+    initMswHandlers(getCoinConfig(scroll).info);
 
     const onSignerConfirmation = getOnSpeculosConfirmation();
     const currencyBridge = buildCurrencyBridge(signerContext);
-    await currencyBridge.preload(blast);
+    await currencyBridge.preload(scroll);
     const accountBridge = buildAccountBridge(signerContext);
     const getAddress = resolver(signerContext);
     const { address } = await getAddress("", {
       path: "44'/60'/0'/0/0",
-      currency: blast,
+      currency: scroll,
       derivationMode: "",
     });
 
-    const scenarioAccount = makeAccount(address, blast);
+    const scenarioAccount = makeAccount(address, scroll);
 
-    const MIM_ON_BLAST = getTokenById(TOKEN_ID);
+    const USDC_ON_SCROLL = getTokenById(TOKEN_ID);
     await callMyDealer({
       provider,
-      drug: MIM_ON_BLAST,
+      drug: USDC_ON_SCROLL,
       junkie: address,
-      dose: ethers.utils.parseUnits("100", MIM_ON_BLAST.units[0].magnitude),
+      dose: ethers.utils.parseUnits("100", USDC_ON_SCROLL.units[0].magnitude),
     });
 
     return {
@@ -129,18 +133,18 @@ export const scenarioBlast: Scenario<EvmTransaction, Account> = {
     await indexBlocks();
   },
   beforeAll: account => {
-    const MIM_ON_BLAST = getTokenById(TOKEN_ID);
+    const USDC_ON_SCROLL = getTokenById(TOKEN_ID);
     expect(account.balance.toFixed()).toBe(ethers.utils.parseEther("10000").toString());
     expect(account.subAccounts?.[0]?.type).toBe("TokenAccount");
     expect(account.subAccounts?.[0]?.balance?.toFixed()).toBe(
-      ethers.utils.parseUnits("100", MIM_ON_BLAST.units[0].magnitude).toString(),
+      ethers.utils.parseUnits("100", USDC_ON_SCROLL.units[0].magnitude).toString(),
     );
   },
   afterAll: account => {
-    const MIM_ON_BLAST = getTokenById(TOKEN_ID);
+    const USDC_ON_SCROLL = getTokenById(TOKEN_ID);
     expect(account.subAccounts?.length).toBe(1);
     expect(account.subAccounts?.[0].balance.toFixed()).toBe(
-      ethers.utils.parseUnits("20", MIM_ON_BLAST.units[0].magnitude).toString(),
+      ethers.utils.parseUnits("20", USDC_ON_SCROLL.units[0].magnitude).toString(),
     );
     expect(account.operations.length).toBe(3);
   },
